requirements.txt
coloredlogs==14.0
junitparser==1.6.2
PyYAML==5.3.1
